Game of Life

+++ src/main/java/com/jt/algo/practice/sword2offer/Sto60.java
@@ -48,10 +48,10 @@ public class Sto60 {
                 }
             }
         }
-        double total = Math.pow((double)6,(double)n);
+        double total = Math.pow(6, n);
         double[] ans = new double[5*n+1];
-        for(int i=n;i<=6*n;i++){
-            ans[i-n]=((double)dp[n][i])/total;
+        for (int i = n; i <= 6 * n; i++) {
+            ans[i - n] = ((double) dp[n][i]) / total;
         }
         return ans;
     }
